+++ src/testResult.ts
@@ -2,7 +2,7 @@ export class TestResult {
     private className: string;
     private method: string;
 
-    public constructor(private _testId: string, private _outcome: string) {
+    public constructor(private _testId: string, private _outcome: string, private _message: string, private _stackTrace: string) {
     }
 
     public get fullName(): string {
@@ -15,6 +15,14 @@ export class TestResult {
 
     public get outcome(): string {
         return this._outcome;
+    }
+
+    public get message(): string {
+        return this._message;
+    }
+
+    public get stackTrace(): string {
+        return this._stackTrace;
     }
 
     public matches(className: string, method: string): boolean {