+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.4 ()
+* [#74](https://github.com/formulahendry/vscode-dotnet-test-explorer/pull/74): Move the view to Test view container
+* [#72](https://github.com/formulahendry/vscode-dotnet-test-explorer/pull/72): Add failed tests to problems view
+* [#70](https://github.com/formulahendry/vscode-dotnet-test-explorer/pull/70): Show spinner for currently running test(s)
+
 ## 0.2.3 (2018-04-27)
 * [#67](https://github.com/formulahendry/vscode-dotnet-test-explorer/pull/67): Option to auto expand tree
 * [#62](https://github.com/formulahendry/vscode-dotnet-test-explorer/pull/62): Better support for xunit theories